requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altgraph==0.17.4
attrs==23.2.0
cffi==1.16.0
cryptography==42.0.4
discord.py==2.3.2
ffmpeg==1.4
frozenlist==1.4.1
idna==3.6
multidict==6.0.5
packaging==23.2
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==1.0.1
tk==0.1.0
yarl==1.9.4
azure-identity
azure-keyvault-secrets